WorkflowExecutionUtils refactoring WorkflowExecutionUtils is split into WorkflowExecutionUtils, WorkflowClientHelper, WorkflowClientLongPollAsyncHelper, WorkflowClientLongPollHelper to encapsulate independent pieces and simplify further testing and maintenance
