+++ temporal-sdk/src/test/java/io/temporal/workflow/SearchAttributesTest.java
@@ -32,7 +32,7 @@ import io.temporal.api.workflowservice.v1.GetWorkflowExecutionHistoryResponse;
 import io.temporal.client.WorkflowClient;
 import io.temporal.client.WorkflowOptions;
 import io.temporal.common.converter.DataConverter;
-import io.temporal.internal.common.WorkflowExecutionUtils;
+import io.temporal.internal.client.WorkflowClientHelper;
 import io.temporal.testing.internal.SDKTestOptions;
 import io.temporal.testing.internal.SDKTestWorkflowRule;
 import io.temporal.workflow.shared.TestMultiArgWorkflowFunctions.TestMultiArgWorkflowImpl;
@@ -99,7 +99,7 @@ public class SearchAttributesTest {
     WorkflowExecution executionF = WorkflowClient.start(stubF::func);
 
     GetWorkflowExecutionHistoryResponse historyResp =
-        WorkflowExecutionUtils.getHistoryPage(
+        WorkflowClientHelper.getHistoryPage(
             testWorkflowRule.getTestEnvironment().getWorkflowService(),
             SDKTestWorkflowRule.NAMESPACE,
             executionF,
